Refactor routers to controller/
