+++ api/controller/food_controller.go
@@ -1,0 +1,53 @@
+package controller
+
+import (
+	"encoding/json"
+	"foodie_manager/db"
+	"io/ioutil"
+
+	"github.com/gin-gonic/gin"
+)
+
+func CreateFood(repo db.FoodRepository) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		var data []byte
+		var err error
+		var food db.Food
+
+		if data, err = ioutil.ReadAll(c.Request.Body); err != nil {
+			c.JSON(400, gin.H{"message": "Error when reading request body."})
+			return
+		}
+
+		if err = json.Unmarshal(data, &food); err != nil {
+			c.JSON(400, gin.H{"message": "Error when parsing request body."})
+			return
+		}
+
+		if err = repo.CreateFood(food); err != nil {
+			c.JSON(500, gin.H{"message": "Something error when creating food."})
+			return
+		}
+		c.JSON(201, gin.H{
+			"message": "Food created",
+		})
+	}
+}
+
+func GetFoodByName(repo db.FoodRepository) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		name := c.Param("name")
+		food := repo.GetFoodByName(name)
+		if len(food.PurchaseDate) == 0 {
+			c.JSON(400, gin.H{"message": "Food not found."})
+			return
+		}
+		c.JSON(200, food)
+	}
+}
+
+func SetupFoodControllers(r *gin.Engine, repo db.FoodRepository) *gin.Engine {
+	r.GET("/foods/:name", GetFoodByName(repo))
+	r.POST("/foods", CreateFood(repo))
+	return r
+}
